backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
Django==5.1.3
django-cors-headers==4.6.0
django-ninja==1.3.0
numpy==2.1.3
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
